Merchant functionality with resourcemanager
Can now purchase weapons from the merchant, still working out a bug with not all items appearing correctly.

+++ Assets/Scripts/Units/Weapons/WeaponManager.cs
@@ -15,7 +15,7 @@ public class WeaponManager : MonoBehaviour
     public float hitChance;
     public float critChance;
     public float requiredAffinity;
-    public float cost;
+    public int cost;
     
     void Start()
     {
